refactor(editor): Extract remaining variables code into features (no-changelog)
