+++ packages/frontend/editor-ui/src/features/environments.ee/components/VariableModal.test.ts
@@ -2,12 +2,12 @@ import { createComponentRenderer } from '@/__tests__/render';
 import { type MockedStore, mockedStore } from '@/__tests__/utils';
 import VariableModal from './VariableModal.vue';
 import { createTestingPinia } from '@pinia/testing';
-import { VARIABLE_MODAL_KEY } from '@/constants';
+import { VARIABLE_MODAL_KEY } from '../environments.constants';
 import { STORES } from '@n8n/stores';
 import userEvent from '@testing-library/user-event';
 import { useUIStore } from '@/stores/ui.store';
-import useEnvironmentsStore from '@/features/environments.ee/environments.store';
-import type { EnvironmentVariable } from '@/features/environments.ee/environments.types';
+import useEnvironmentsStore from '../environments.store';
+import type { EnvironmentVariable } from '../environments.types';
 import { useProjectsStore } from '@/features/projects/projects.store';
 
 vi.mock('vue-router', () => ({
