+++ packages/frontend/editor-ui/src/features/environments.ee/environments.constants.ts
@@ -1,0 +1,1 @@
+export const VARIABLE_MODAL_KEY = 'variableModal';
