+++ packages/frontend/editor-ui/src/stores/ui.store.ts
@@ -32,9 +32,9 @@ import {
 	EXPERIMENT_TEMPLATE_RECO_V2_KEY,
 	CONFIRM_PASSWORD_MODAL_KEY,
 	EXPERIMENT_TEMPLATE_RECO_V3_KEY,
-	VARIABLE_MODAL_KEY,
 } from '@/constants';
 import { DEBUG_PAYWALL_MODAL_KEY } from '@/features/executions/executions.constants';
+import { VARIABLE_MODAL_KEY } from '@/features/environments.ee/environments.constants';
 import { COMMUNITY_PLUS_ENROLLMENT_MODAL } from '@/features/usage/usage.constants';
 import {
 	CREDENTIAL_EDIT_MODAL_KEY,
